requirements.txt
fastapi
uvicorn[standard]

strawberry-graphql[fastapi]

sqlalchemy>=2.0
asyncpg
psycopg2-binary

python-dotenv
alembic
